[REF] analytic_tag_dimension: Follow OCA guidelines of one file per main model
