+++ analytic_tag_dimension/models/__init__.py
@@ -1,2 +1,6 @@
-from . import analytic
+from . import analytic_dimension_line  # this one should go first
+from . import account_analytic_dimension
+from . import account_analytic_line
+from . import account_analytic_tag
 from . import account_invoice_report
+from . import account_move_line
